Varous updates to adopt QCache to Python 3.
24 test cases are still failing but will be left for now.

The only real advantage of moving over to Python 3 at this stage would
be the improved pickling support (much speedier) in protocol version 3.

Will make arcitectural updates to reduce tail latencies due to
serialization/deserialization in the same thread as querying
and to improve the case for connection pooling before returning
to potential Python 3 migration. Some CSV related operations
are slower in Python 3 compared to Python 2 and there seems to
be an issue with numexpr and unicode strings that prevent filtering
on string values for now...